app/requirements.txt
flask
numpy
pandas
scikit-learn
mysql-connector-python
